rename: api 관련 디렉토리 폴더명 변경
