+++ frontend/src/pages/NewPin.tsx
@@ -4,10 +4,10 @@ import Flex from '../components/common/Flex';
 import Space from '../components/common/Space';
 import Button from '../components/common/Button';
 import Textarea from '../components/common/Textarea';
-import { postApi } from '../utils/postApi';
+import { postApi } from '../apis/postApi';
 import { FormEvent, useContext, useEffect, useRef, useState } from 'react';
-import { getApi } from '../utils/getApi';
-import { getAddress } from '../utils/getAddress';
+import { getApi } from '../apis/getApi';
+import { getAddress } from '../apis/getAddress';
 import { TopicType } from '../types/Topic';
 import useNavigator from '../hooks/useNavigator';
 import { NewPinValuesType } from '../types/FormValues';
